rename feedback to inquiry
