--- backend/src/controllers/feedback.controller.ts
@@ -1,20 +1,0 @@
-import { Request, Response } from "express";
-import catchErrors from "../utils/catchErrors";
-import { OK } from "../constants/http";
-import { sendFeedbackSchema } from "../schemas/feedback.schema";
-import { sendFeedback } from "../services/feedback.service";
-
-export const sendFeedbackHandler = catchErrors(
-  async (request: Request, response: Response) => {
-    const data = sendFeedbackSchema.parse(request.body);
-
-    await sendFeedback(data);
-
-    response.status(OK).json({
-      status: "success",
-      data: {
-        message: "Feedback sent successfully",
-      },
-    });
-  }
-);
