+++ backend/src/controllers/inquiry.controller.ts
@@ -1,0 +1,20 @@
+import { Request, Response } from "express";
+import catchErrors from "../utils/catchErrors";
+import { OK } from "../constants/http";
+import { sendFeedbackSchema } from "../schemas/feedback.schema";
+import { sendFeedback } from "../services/inquiry.service";
+
+export const sendInquiryHandler = catchErrors(
+  async (request: Request, response: Response) => {
+    const data = sendFeedbackSchema.parse(request.body);
+
+    await sendFeedback(data);
+
+    response.status(OK).json({
+      status: "success",
+      data: {
+        message: "Feedback sent successfully",
+      },
+    });
+  }
+);
