+++ backend/src/routes/inquiry.routes.ts
@@ -1,0 +1,8 @@
+import { Router } from "express";
+import { sendInquiryHandler } from "../controllers/inquiry.controller";
+
+const router = Router();
+
+router.post("", sendInquiryHandler);
+
+export default router;
