+++ backend/src/routes/routes.ts
@@ -2,10 +2,11 @@ import { Router, Request, Response } from "express";
 import { OK } from "../constants/http";
 import authRoutes from "./auth.routes";
 import userRoutes from "./user.routes";
-import feedbackRoutes from "./feedback.routes";
+import inquiryRoutes from "./inquiry.routes";
 import serviceRoutes from "./service.routes";
 import sessionRoutes from "./session.routes";
 import bookingRoutes from "./booking.routes";
+import feedbackRoutes from "./feedback.routes"
 import { authenticate } from "../middlewares/auth.middleware";
 import { upload } from "../middlewares/upload";
 
@@ -24,7 +25,8 @@ router.use("/auth", authRoutes);
 router.use("/sessions", sessionRoutes);
 router.use("/users", authenticate, userRoutes);
 router.use("/services", serviceRoutes);
-router.use("/feedbacks", feedbackRoutes);
+router.use("/inquiry", inquiryRoutes);
+router.use("/feedback", feedbackRoutes)
 router.use("/services", serviceRoutes);
 router.use("/bookings", bookingRoutes);
